+++ blacksheep/server/authentication/cookie.py
@@ -94,13 +94,13 @@ class CookieAuthentication(AuthenticationHandler):
         """
         Sets the user context, when user's data was parsed and validated from a cookie.
         """
-        context.identity = Identity(data, self.auth_scheme)
+        context.user = Identity(data, self.auth_scheme)
 
     async def authenticate(self, context: Request) -> Optional[Identity]:
         cookie = context.get_cookie(self.cookie_name)
 
         if cookie is None:
-            context.identity = Identity({})
+            context.user = Identity({})
         else:
             try:
                 value = self.serializer.loads(cookie)
@@ -109,7 +109,7 @@ class CookieAuthentication(AuthenticationHandler):
                     "Cookie authentication failed (%s), invalid signature.",
                     self.cookie_name,
                 )
-                context.identity = Identity({})
+                context.user = Identity({})
             else:
                 self.set_user_context(context, value)
         return None