+++ itests/test_auth_oidc.py
@@ -2,6 +2,8 @@
 This module is under integration tests folder because it uses a partially faked
 authorization server in a running Flask server.
 """
+import json
+import re
 from datetime import datetime
 from typing import Any, List, Optional, Tuple
 from urllib.parse import parse_qs, urlencode
@@ -10,14 +12,19 @@ import pytest
 from guardpost import Identity, Policy
 from guardpost.common import AuthenticatedRequirement
 
+from blacksheep.contents import Content
 from blacksheep.cookies import parse_cookie
 from blacksheep.exceptions import BadRequest, Unauthorized
 from blacksheep.messages import Request, Response
 from blacksheep.server.application import Application
 from blacksheep.server.asgi import incoming_request
 from blacksheep.server.authentication.cookie import CookieAuthentication
+from blacksheep.server.authentication.jwt import JWTBearerAuthentication
 from blacksheep.server.authentication.oidc import (
+    CookiesOpenIDTokensHandler,
     CookiesTokensStore,
+    IDToken,
+    JWTOpenIDTokensHandler,
     MissingClientSecretSettingError,
     OpenIDConfiguration,
     OpenIDConnectConfigurationError,
@@ -28,13 +35,14 @@ from blacksheep.server.authentication.oidc import (
     OpenIDSettings,
     ParametersBuilder,
     TokenResponse,
+    TokensStore,
     use_openid_connect,
 )
 from blacksheep.testing.helpers import get_example_scope
 from blacksheep.testing.messages import MockReceive, MockSend
 from blacksheep.url import URL
 from blacksheep.utils.aio import FailedRequestError
-from tests.test_auth import get_access_token
+from tests.test_auth import get_token
 from tests.test_auth_cookie import get_auth_cookie
 from tests.utils.application import FakeApplication
 
@@ -43,17 +51,110 @@ from .client_fixtures import *  # NoQA
 MOCKED_AUTHORITY = "http://127.0.0.1:44777/oidc"
 
 
+try:
+    from unittest.mock import AsyncMock
+except ImportError:
+    # Python 3.7
+    from mock import AsyncMock
+
+
 @pytest.fixture
 def app():
     return FakeApplication()
 
 
-def configure_test_oidc_implicit_id_token(app: Application):
+def _mock_token_endpoint(oidc, oidc_configuration, id_token_claims):
+    # mock handling of the remote OIDC server
+    async def mocked_post_form(url: str, data: Any):
+        assert url == oidc_configuration.token_endpoint
+
+        access_token_claims = {
+            "sub": "4534224f-546f-401f-9cab-067b0b2b9abb",
+            "aud": "------------------------------------",
+            "scp": "read:todos",
+            "iss": oidc_configuration.issuer,
+        }
+
+        return {
+            "id_token": get_token("0", id_token_claims),
+            "access_token": get_token("0", access_token_claims),
+            "refresh_token": "00000000-0000-0000-0000-000000000000",
+        }
+
+    oidc._http_handler.post_form = mocked_post_form
+
+
+class FakeTokensStore(TokensStore):
+    def __init__(
+        self,
+        access_token: Optional[str] = None,
+        refresh_token: Optional[str] = None,
+    ) -> None:
+        super().__init__()
+        self._access_token = access_token
+        self._refresh_token = refresh_token
+
+    async def store_tokens(
+        self,
+        request: Request,
+        response: Response,
+        access_token: str,
+        refresh_token: Optional[str],
+        expires: Optional[datetime] = None,
+    ):
+        """
+        Applies a strategy to store an access token and an optional refresh token for
+        the given request and response.
+        """
+        self._access_token = access_token
+        self._refresh_token = refresh_token
+
+    async def unset_tokens(self, request: Request):
+        """
+        Optional method, to unset access tokens upon sign-out.
+        """
+        self._access_token = None
+        self._refresh_token = None
+
+    async def restore_tokens(self, request: Request) -> None:
+        """
+        Applies a strategy to restore an access token and an optional refresh token for
+        the given request.
+        """
+        assert request.identity is not None
+        request.identity.access_token = self._access_token
+        request.identity.refresh_token = self._refresh_token
+
+
+def configure_test_oidc_cookie_auth_id_token(
+    app: Application, secret: Optional[str] = None
+):
     return use_openid_connect(
         app,
         OpenIDSettings(
             client_id="067cee45-faf3-4c75-9fef-09f050bcc3ae",
+            client_secret=secret,
             authority=MOCKED_AUTHORITY,
+        ),
+    )
+
+
+def configure_test_oidc_jwt_auth_id_token(
+    app: Application, secret: Optional[str] = None
+):
+    CLIENT_ID = "067cee45-faf3-4c75-9fef-09f050bcc3ae"
+    return use_openid_connect(
+        app,
+        OpenIDSettings(
+            client_id=CLIENT_ID,
+            client_secret=secret,
+            authority=MOCKED_AUTHORITY,
+        ),
+        auth_handler=JWTOpenIDTokensHandler(
+            JWTBearerAuthentication(
+                authority=MOCKED_AUTHORITY,
+                valid_audiences=[CLIENT_ID],
+            )
         ),
     )
 
@@ -98,7 +199,19 @@ def assert_redirect_to_sign_in(response: Optional[Response], has_secret: bool = 
 
 @pytest.mark.asyncio
 async def test_oidc_handler_redirect(app: FakeApplication):
-    oidc = configure_test_oidc_implicit_id_token(app)
+    oidc = configure_test_oidc_cookie_auth_id_token(app)
+    assert isinstance(oidc, OpenIDConnectHandler)
+
+    await app(
+        get_example_scope("GET", oidc.settings.entry_path), MockReceive(), MockSend()
+    )
+
+    assert_redirect_to_sign_in(app.response)
+
+
+@pytest.mark.asyncio
+async def test_oidc_handler_redirect_with_jwt_handler(app: FakeApplication):
+    oidc = configure_test_oidc_jwt_auth_id_token(app)
     assert isinstance(oidc, OpenIDConnectHandler)
 
     await app(
@@ -121,8 +234,12 @@ async def test_oidc_handler_redirect_with_secret(app: FakeApplication):
 
 
 @pytest.mark.asyncio
-async def test_oidc_handler_auth_post_id_token(app: FakeApplication):
-    oidc = configure_test_oidc_implicit_id_token(app)
+async def test_oidc_handler_cookie_auth_post_id_token(app: FakeApplication):
+    """
+    Tests the response from the built-in CookiesOpenIDTokensHandler handler after a
+    successful sign-in.
+    """
+    oidc = configure_test_oidc_cookie_auth_id_token(app)
     assert isinstance(oidc, OpenIDConnectHandler)
 
     oidc_configuration = await oidc.get_openid_configuration()
@@ -141,7 +258,7 @@ async def test_oidc_handler_auth_post_id_token(app: FakeApplication):
         assert parsed_token == claims
 
     # arrange an id_token set by the remote auth server
-    id_token = get_access_token("0", claims)
+    id_token = get_token("0", claims)
     content = urlencode({"id_token": id_token}).encode()
 
     await app(
@@ -166,8 +283,245 @@ async def test_oidc_handler_auth_post_id_token(app: FakeApplication):
     cookie = parse_cookie(cookie_value)
 
     # the auth_handler can parse the cookie value:
-    parsed_cookie_value = oidc._auth_handler.serializer.loads(cookie.value)
+    assert isinstance(oidc.auth_handler, CookiesOpenIDTokensHandler)
+    parsed_cookie_value = oidc.auth_handler.auth_handler.serializer.loads(cookie.value)
     assert parsed_cookie_value == claims
+
+
+@pytest.mark.asyncio
+async def test_oidc_handler_jwt_auth_post_id_token(app: FakeApplication):
+    """
+    Tests the response from the built-in JWTOpenIDTokensHandler handler after a
+    successful sign-in.
+    """
+    oidc = configure_test_oidc_jwt_auth_id_token(app)
+    assert isinstance(oidc, OpenIDConnectHandler)
+
+    oidc_configuration = await oidc.get_openid_configuration()
+
+    called = False
+    claims = {
+        "aud": oidc.settings.client_id,
+        "iss": oidc_configuration.issuer,
+        "sub": "4534224f-546f-401f-9cab-067b0b2b9abb",
+    }
+
+    @oidc.events.on_id_token_validated
+    async def on_id_token_validated(context, parsed_token):
+        nonlocal called
+        called = True
+        assert parsed_token == claims
+
+    # arrange an id_token set by the remote auth server
+    id_token = get_token("0", claims)
+    content = urlencode({"id_token": id_token}).encode()
+
+    await app(
+        get_example_scope(
+            "POST",
+            oidc.settings.callback_path,
+            form_extra_headers(content),
+        ),
+        MockReceive([content]),
+        MockSend(),
+    )
+
+    response = app.response
+    assert response is not None
+
+    assert called is True
+    assert response.status == 200
+    html = await response.text()
+
+    assert 'sessionStorage.setItem("ID_TOKEN",' in html
+    assert 'sessionStorage.setItem("ACCESS_TOKEN", "");' in html
+    assert 'sessionStorage.setItem("REFRESH_TOKEN", "");' in html
+    match = re.search(r'"ID_TOKEN",\s"([^\"]+)\"', html)
+    assert match
+
+    parsed_id_token = IDToken.from_trusted_token(match.group(1))
+    assert parsed_id_token.data == claims
+
+
+@pytest.mark.asyncio
+async def test_oidc_handler_jwt_refresh_token(app: FakeApplication):
+    """
+    Tests handling of refresh tokens using the JWT handler.
+    """
+    oidc = configure_test_oidc_jwt_auth_id_token(app, "EXAMPLE")
+
+    await app.start()
+    oidc_configuration = await oidc.get_openid_configuration()
+
+    claims = {
+        "aud": oidc.settings.client_id,
+        "iss": oidc_configuration.issuer,
+        "sub": "4534224f-546f-401f-9cab-067b0b2b9abb",
+    }
+
+    auth_handler = oidc.auth_handler
+    assert isinstance(auth_handler, JWTOpenIDTokensHandler)
+
+    _mock_token_endpoint(oidc, oidc_configuration, claims)
+
+    # call to obtain fresh tokens
+    await app(
+        get_example_scope(
+            "POST",
+            oidc.settings.refresh_token_path,
+            extra_headers={
+                "X-Refresh-Token": auth_handler.protect_refresh_token("TEST")
+            },
+        ),
+        MockReceive(),
+        MockSend(),
+    )
+
+    response = app.response
+    assert response is not None
+
+    assert response.status == 200
+    data = await response.json()
+
+    assert "id_token" in data
+    assert "access_token" in data
+    assert "refresh_token" in data
+
+
+@pytest.mark.asyncio
+async def test_oidc_handler_jwt_refresh_token_invalid_token(app: FakeApplication):
+    """
+    Tests that the JWT handler ignores invalid refresh tokens.
+    """
+    oidc = configure_test_oidc_jwt_auth_id_token(app, "EXAMPLE")
+
+    await app.start()
+    oidc_configuration = await oidc.get_openid_configuration()
+
+    claims = {
+        "aud": oidc.settings.client_id,
+        "iss": oidc_configuration.issuer,
+        "sub": "4534224f-546f-401f-9cab-067b0b2b9abb",
+    }
+
+    auth_handler = oidc.auth_handler
+    assert isinstance(auth_handler, JWTOpenIDTokensHandler)
+
+    _mock_token_endpoint(oidc, oidc_configuration, claims)
+
+    # call to obtain fresh tokens
+    await app(
+        get_example_scope(
+            "POST",
+            oidc.settings.refresh_token_path,
+            extra_headers={"X-Refresh-Token": "INVALID_TOKEN"},
+        ),
+        MockReceive(),
+        MockSend(),
+    )
+
+    response = app.response
+    assert response is not None
+    assert response.status == 400
+
+    text = await response.text()
+    assert "Missing refresh_token" in text
+
+
+@pytest.mark.asyncio
+async def test_oidc_handler_cookie_refresh_token(app: FakeApplication):
+    """
+    Tests handling of refresh tokens using the JWT handler.
+    """
+    oidc = configure_test_oidc_cookie_auth_id_token(app, secret="TEST_EXAMPLE")
+    oidc.auth_handler.tokens_store = FakeTokensStore(refresh_token="TEST")  # type: ignore
+
+    await app.start()
+    oidc_configuration = await oidc.get_openid_configuration()
+
+    claims = {
+        "aud": oidc.settings.client_id,
+        "iss": oidc_configuration.issuer,
+        "sub": "4534224f-546f-401f-9cab-067b0b2b9abb",
+    }
+
+    _mock_token_endpoint(oidc, oidc_configuration, claims)
+
+    # call to obtain fresh tokens
+    await app(
+        get_example_scope("POST", oidc.settings.refresh_token_path),
+        MockReceive(),
+        MockSend(),
+    )
+
+    response = app.response
+    assert response is not None
+
+    assert response.status == 200
+
+    cookie_value = response.headers.get_single(b"set-cookie")
+
+    assert cookie_value is not None
+    cookie = parse_cookie(cookie_value)
+
+    # the auth_handler can parse the cookie value:
+    assert isinstance(oidc.auth_handler, CookiesOpenIDTokensHandler)
+    parsed_cookie_value = oidc.auth_handler.auth_handler.serializer.loads(cookie.value)
+    assert parsed_cookie_value == claims
+
+
+@pytest.mark.asyncio
+async def test_oidc_handler_refresh_token_missing_user_context(app: FakeApplication):
+    oidc = configure_test_oidc_cookie_auth_id_token(app, secret="TEST_EXAMPLE")
+    oidc.auth_handler.tokens_store = FakeTokensStore(refresh_token=None)  # type: ignore
+
+    await app(
+        get_example_scope("POST", oidc.settings.refresh_token_path),
+        MockReceive(),
+        MockSend(),
+    )
+
+    response = app.response
+    assert response is not None
+
+    assert response.status == 400
+
+    text = await response.text()
+    assert "Missing user" in text
+
+
+@pytest.mark.asyncio
+async def test_oidc_handler_refresh_token_missing_refresh_token_context(
+    app: FakeApplication,
+):
+    oidc = configure_test_oidc_cookie_auth_id_token(app, secret="TEST_EXAMPLE")
+    oidc.auth_handler.tokens_store = FakeTokensStore(refresh_token=None)  # type: ignore
+
+    await app.start()
+    oidc_configuration = await oidc.get_openid_configuration()
+
+    claims = {
+        "aud": oidc.settings.client_id,
+        "iss": oidc_configuration.issuer,
+        "sub": "4534224f-546f-401f-9cab-067b0b2b9abb",
+    }
+
+    _mock_token_endpoint(oidc, oidc_configuration, claims)
+
+    # call to obtain fresh tokens
+    await app(
+        get_example_scope("POST", oidc.settings.refresh_token_path),
+        MockReceive(),
+        MockSend(),
+    )
+
+    response = app.response
+    assert response is not None
+
+    assert response.status == 400
+
+    text = await response.text()
+    assert "Missing refresh_token" in text
 
 
 @pytest.mark.asyncio
@@ -197,7 +551,7 @@ async def test_oidc_handler_auth_post_id_token_code_1(app: FakeApplication):
         }
 
         return {
-            "access_token": get_access_token("0", access_token_claims),
+            "access_token": get_token("0", access_token_claims),
             "refresh_token": "00000000-0000-0000-0000-000000000000",
         }
 
@@ -218,7 +572,7 @@ async def test_oidc_handler_auth_post_id_token_code_1(app: FakeApplication):
         called = True
 
     # arrange an id_token set by the remote auth server
-    id_token = get_access_token("0", id_token_claims)
+    id_token = get_token("0", id_token_claims)
     content = urlencode({"id_token": id_token, "code": "xxx"}).encode()
 
     await app(
@@ -243,7 +597,8 @@ async def test_oidc_handler_auth_post_id_token_code_1(app: FakeApplication):
     cookie = parse_cookie(cookie_value)
 
     # the auth_handler can parse the cookie value:
-    parsed_cookie_value = oidc._auth_handler.serializer.loads(cookie.value)
+    assert isinstance(oidc.auth_handler, CookiesOpenIDTokensHandler)
+    parsed_cookie_value = oidc.auth_handler.auth_handler.serializer.loads(cookie.value)
     assert parsed_cookie_value == id_token_claims
 
 
@@ -282,8 +637,8 @@ async def test_oidc_handler_auth_post_id_token_code_2(app: FakeApplication):
         }
 
         return {
-            "id_token": get_access_token("0", id_token_claims),
-            "access_token": get_access_token("0", access_token_claims),
+            "id_token": get_token("0", id_token_claims),
+            "access_token": get_token("0", access_token_claims),
             "refresh_token": "00000000-0000-0000-0000-000000000000",
         }
 
@@ -323,7 +678,8 @@ async def test_oidc_handler_auth_post_id_token_code_2(app: FakeApplication):
     cookie = parse_cookie(cookie_value)
 
     # the auth_handler can parse the cookie value:
-    parsed_cookie_value = oidc._auth_handler.serializer.loads(cookie.value)
+    assert isinstance(oidc.auth_handler, CookiesOpenIDTokensHandler)
+    parsed_cookie_value = oidc.auth_handler.auth_handler.serializer.loads(cookie.value)
     assert parsed_cookie_value == id_token_claims
 
 
@@ -343,7 +699,7 @@ async def test_redirect_state_includes_original_path(
     Tests the ability to redirect the user to the original path that was requested
     before a redirect to the OIDC sign-in page, using the state parameter.
     """
-    oidc = configure_test_oidc_implicit_id_token(app)
+    oidc = configure_test_oidc_cookie_auth_id_token(app)
 
     oidc_configuration = await oidc.get_openid_configuration()
 
@@ -392,7 +748,7 @@ async def test_redirect_state_includes_original_path(
         "sub": "4534224f-546f-401f-9cab-067b0b2b9abb",
         "nonce": parsed_state.get("nonce"),
     }
-    id_token = get_access_token("0", id_token_claims)
+    id_token = get_token("0", id_token_claims)
     content = urlencode({"id_token": id_token, "state": state_value}).encode()
 
     await app(
@@ -420,7 +776,7 @@ async def test_raises_for_nonce_mismatch(app: FakeApplication):
     Tests the ability to redirect the user to the original path that was requested
     before a redirect to the OIDC sign-in page, using the state parameter.
     """
-    oidc = configure_test_oidc_implicit_id_token(app)
+    oidc = configure_test_oidc_cookie_auth_id_token(app)
 
     oidc_configuration = await oidc.get_openid_configuration()
 
@@ -457,7 +813,7 @@ async def test_raises_for_nonce_mismatch(app: FakeApplication):
         "sub": "4534224f-546f-401f-9cab-067b0b2b9abb",
         "nonce": "this will not match",
     }
-    id_token = get_access_token("0", id_token_claims)
+    id_token = get_token("0", id_token_claims)
     content = urlencode({"id_token": id_token, "state": state_value}).encode()
     scope = get_example_scope(
         "POST",
@@ -474,7 +830,7 @@ async def test_raises_for_nonce_mismatch(app: FakeApplication):
 
 @pytest.mark.asyncio
 async def test_raises_for_missing_data(app: FakeApplication):
-    oidc = configure_test_oidc_implicit_id_token(app)
+    oidc = configure_test_oidc_cookie_auth_id_token(app)
 
     content = urlencode({"invalid": 1}).encode()
     scope = get_example_scope(
@@ -555,7 +911,7 @@ async def test_raises_for_invalid_id_token(app: FakeApplication):
     """
     Tests handling of forged id_token.
     """
-    oidc = configure_test_oidc_implicit_id_token(app)
+    oidc = configure_test_oidc_cookie_auth_id_token(app)
 
     oidc_configuration = await oidc.get_openid_configuration()
 
@@ -568,8 +924,8 @@ async def test_raises_for_invalid_id_token(app: FakeApplication):
     }
 
     for forged_id_token in [
-        get_access_token("foreign", id_token_claims),
-        get_access_token("foreign", id_token_claims, fake_kid="0"),
+        get_token("foreign", id_token_claims),
+        get_token("foreign", id_token_claims, fake_kid="0"),
     ]:
         content = urlencode({"id_token": forged_id_token}).encode()
         scope = get_example_scope(
@@ -654,7 +1010,7 @@ async def test_oidc_handler_with_secret_and_audience_no_id_token(
         }
 
         return {
-            "access_token": get_access_token("0", access_token_claims),
+            "access_token": get_token("0", access_token_claims),
             "refresh_token": "00000000-0000-0000-0000-000000000000",
         }
 
@@ -682,7 +1038,8 @@ async def test_oidc_handler_with_secret_and_audience_no_id_token(
     assert cookie_value is not None
     cookie = parse_cookie(cookie_value)
 
-    parsed_cookie_value = oidc._auth_handler.serializer.loads(cookie.value)
+    assert isinstance(oidc.auth_handler, CookiesOpenIDTokensHandler)
+    parsed_cookie_value = oidc.auth_handler.auth_handler.serializer.loads(cookie.value)
     assert parsed_cookie_value == {}
 
 
@@ -690,9 +1047,10 @@ async def test_oidc_handler_with_secret_and_audience_no_id_token(
 async def test_oidc_handler_auth_post_id_token_code_3(app: FakeApplication):
     """With token store"""
     oidc = configure_test_oidc_with_secret(app)
-    oidc.tokens_store = CookiesTokensStore(oidc.settings.scheme_name)
+    assert isinstance(oidc.auth_handler, CookiesOpenIDTokensHandler)
+    oidc.auth_handler.tokens_store = CookiesTokensStore(oidc.settings.scheme_name)
 
-    assert oidc.tokens_store.scheme_name == oidc.settings.scheme_name
+    assert oidc.auth_handler.tokens_store.scheme_name == oidc.settings.scheme_name
 
     # Note: mock the HTTPHandler to simulate a real authorization server that
     # handles a valid request to the token endpoint
@@ -713,8 +1071,8 @@ async def test_oidc_handler_auth_post_id_token_code_3(app: FakeApplication):
         }
 
         return {
-            "id_token": get_access_token("0", id_token_claims),
-            "access_token": get_access_token("0", access_token_claims),
+            "id_token": get_token("0", id_token_claims),
+            "access_token": get_token("0", access_token_claims),
             "refresh_token": "00000000-0000-0000-0000-000000000000",
         }
 
@@ -754,6 +1112,7 @@ async def test_oidc_handler_auth_post_id_token_code_3(app: FakeApplication):
 async def test_cookies_tokens_store_restoring_context(
     app: FakeApplication,
 ):
+    tokens_store = CookiesTokensStore()
     oidc = use_openid_connect(
         app,
         OpenIDSettings(
@@ -761,13 +1120,16 @@ async def test_cookies_tokens_store_restoring_context(
             client_secret="JUST_AN_EXAMPLE",
             authority=MOCKED_AUTHORITY,
         ),
-        tokens_store=CookiesTokensStore(),
+        auth_handler=CookiesOpenIDTokensHandler(
+            CookieAuthentication(
+                cookie_name="OpenIDConnect".lower(), auth_scheme="OpenIDConnect"
+            ),
+            tokens_store=tokens_store,
+        ),
     )
-    assert oidc.tokens_store in app.middlewares
 
-    assert isinstance(oidc.tokens_store, CookiesTokensStore)
-    access_token = oidc.tokens_store.serializer.dumps("secret")
-    refresh_token = oidc.tokens_store.serializer.dumps("secret-refresh")
+    access_token = tokens_store.serializer.dumps("secret")
+    refresh_token = tokens_store.serializer.dumps("secret-refresh")
 
     assert isinstance(access_token, str)
     assert isinstance(refresh_token, str)
@@ -781,29 +1143,11 @@ async def test_cookies_tokens_store_restoring_context(
         )
     )
 
-    request.identity = Identity({})
-    await oidc.tokens_store.restore_tokens(request)
+    request.user = Identity({})
+    await tokens_store.restore_tokens(request)
 
-    assert request.identity.access_token == "secret"
-    assert request.identity.refresh_token == "secret-refresh"
-
-    request = incoming_request(
-        get_example_scope(
-            "GET",
-            "/",
-            cookies={scheme + ".at": access_token, scheme + ".rt": refresh_token},
-        )
-    )
-
-    request.identity = Identity({})
-
-    async def handler(*args):
-        pass
-
-    await oidc.tokens_store(request, handler)  # middleware way
-
-    assert request.identity.access_token == "secret"
-    assert request.identity.refresh_token == "secret-refresh"
+    assert request.user.access_token == "secret"
+    assert request.user.refresh_token == "secret-refresh"
 
 
 @pytest.mark.asyncio
@@ -811,7 +1155,9 @@ async def test_cookies_tokens_store_discards_invalid_tokens(
     app: FakeApplication,
 ):
     oidc = configure_test_oidc_with_secret(app)
-    oidc.tokens_store = CookiesTokensStore(oidc.settings.scheme_name)
+    assert isinstance(oidc.auth_handler, CookiesOpenIDTokensHandler)
+    tokens_store = CookiesTokensStore(oidc.settings.scheme_name)
+    oidc.auth_handler.tokens_store = tokens_store
 
     access_token = "invalid"
     refresh_token = "invalid"
@@ -825,11 +1171,11 @@ async def test_cookies_tokens_store_discards_invalid_tokens(
         )
     )
 
-    request.identity = Identity({})
-    await oidc.tokens_store.restore_tokens(request)
+    request.user = Identity({})
+    await tokens_store.restore_tokens(request)
 
-    assert getattr(request.identity, "access_token", None) is None
-    assert getattr(request.identity, "refresh_token", None) is None
+    assert getattr(request.user, "access_token", None) is None
+    assert getattr(request.user, "refresh_token", None) is None
 
 
 @pytest.mark.asyncio
@@ -837,7 +1183,8 @@ async def test_cookies_tokens_store_handle_missing_cookies(
     app: FakeApplication,
 ):
     oidc = configure_test_oidc_with_secret(app)
-    oidc.tokens_store = CookiesTokensStore(oidc.settings.scheme_name)
+    tokens_store = CookiesTokensStore(oidc.settings.scheme_name)
+    oidc.auth_handler.tokens_store = tokens_store
 
     request = incoming_request(
         get_example_scope(
@@ -846,16 +1193,16 @@ async def test_cookies_tokens_store_handle_missing_cookies(
         )
     )
 
-    request.identity = Identity({})
-    await oidc.tokens_store.restore_tokens(request)
+    request.user = Identity({})
+    await tokens_store.restore_tokens(request)
 
-    assert getattr(request.identity, "access_token", None) is None
-    assert getattr(request.identity, "refresh_token", None) is None
+    assert getattr(request.user, "access_token", None) is None
+    assert getattr(request.user, "refresh_token", None) is None
 
 
 @pytest.mark.asyncio
 async def test_oidc_handler_auth_post_error(app: FakeApplication):
-    oidc = configure_test_oidc_implicit_id_token(app)
+    oidc = configure_test_oidc_cookie_auth_id_token(app)
     assert isinstance(oidc, OpenIDConnectHandler)
 
     called = False
@@ -889,7 +1236,7 @@ async def test_oidc_handler_auth_post_error(app: FakeApplication):
 async def test_oidc_handler_handling_request_without_host_header(
     app: FakeApplication,
 ):
-    oidc = configure_test_oidc_implicit_id_token(app)
+    oidc = configure_test_oidc_cookie_auth_id_token(app)
 
     scope = get_example_scope("GET", oidc.settings.entry_path)
     scope["headers"] = []
@@ -909,7 +1256,7 @@ async def test_oidc_handler_handling_request_without_host_header(
 async def test_default_openid_connect_handler_redirects_unauthenticated_users(
     app: FakeApplication,
 ):
-    oidc = configure_test_oidc_implicit_id_token(app)
+    oidc = configure_test_oidc_cookie_auth_id_token(app)
 
     app.use_authorization().with_default_policy(
         Policy("authenticated", AuthenticatedRequirement())
@@ -930,6 +1277,7 @@ async def test_default_openid_connect_handler_redirects_unauthenticated_users(
 
     # if the user is authenticated, but not authorized to do something,
     # then the server should not redirect to the sign-in page
+    assert isinstance(oidc.auth_handler, CookiesOpenIDTokensHandler)
     await app(
         get_example_scope(
             "GET",
@@ -938,7 +1286,8 @@ async def test_default_openid_connect_handler_redirects_unauthenticated_users(
                 (
                     b"cookie",
                     get_auth_cookie(
-                        oidc._auth_handler, {"id": 1, "email": "example@neoteroi.dev"}
+                        oidc.auth_handler.auth_handler,
+                        {"id": 1, "email": "example@neoteroi.dev"},
                     ).encode(),
                 )
             ],
@@ -956,7 +1305,7 @@ async def test_default_openid_connect_handler_redirects_unauthenticated_users(
 async def test_oidc_handler_logout_endpoint(
     app: FakeApplication,
 ):
-    oidc = configure_test_oidc_implicit_id_token(app)
+    oidc = configure_test_oidc_cookie_auth_id_token(app)
 
     await app.start()
     await app(
@@ -1099,7 +1448,7 @@ def test_token_response_class():
 
 def test_raises_for_missing_authority_and_discovery_endpoint():
     handler = OpenIDConnectHandler(
-        OpenIDSettings(client_id="1"), auth_handler=CookieAuthentication()
+        OpenIDSettings(client_id="1"), auth_handler=CookiesOpenIDTokensHandler()
     )
 
     with pytest.raises(OpenIDConnectConfigurationError):
@@ -1110,7 +1459,7 @@ def test_raises_for_missing_authority_and_discovery_endpoint():
 async def test_raises_for_failed_request_to_fetch_openid_configuration():
     handler = OpenIDConnectHandler(
         OpenIDSettings(client_id="1", discovery_endpoint="http://localhost:44123"),
-        auth_handler=CookieAuthentication(),
+        auth_handler=CookiesOpenIDTokensHandler(),
     )
 
     with pytest.raises(OpenIDConnectRequestError):
@@ -1119,7 +1468,7 @@ async def test_raises_for_failed_request_to_fetch_openid_configuration():
 
 @pytest.mark.asyncio
 async def test_oidc_handler_auth_post_without_input(app: FakeApplication):
-    oidc = configure_test_oidc_implicit_id_token(app)
+    oidc = configure_test_oidc_cookie_auth_id_token(app)
     assert isinstance(oidc, OpenIDConnectHandler)
 
     await app(
@@ -1135,3 +1484,87 @@ async def test_oidc_handler_auth_post_without_input(app: FakeApplication):
     assert response is not None
 
     assert response.status == 202  # accepted
+
+
+@pytest.mark.asyncio
+async def test_logout_cookie_handler_unset_tokens():
+    tokens_store = FakeTokensStore()
+    tokens_store.unset_tokens = AsyncMock()
+    handler = CookiesOpenIDTokensHandler(tokens_store=tokens_store)
+    logout_request = Request("GET", b"/logout", [])
+    await handler.get_logout_response(logout_request, "/")
+
+    assert tokens_store.unset_tokens.call_count == 1
+    assert tokens_store.unset_tokens.call_args == ((logout_request,),)
+
+
+@pytest.mark.asyncio
+async def test_jwt_handler_get_logout_response():
+    handler = JWTOpenIDTokensHandler(
+        JWTBearerAuthentication(
+            authority=MOCKED_AUTHORITY,
+            valid_audiences=["NULL"],
+        )
+    )
+    logout_request = Request("GET", b"/logout", [])
+    response = await handler.get_logout_response(logout_request, "/")
+
+    html = response.content.body.decode("utf8")  # type: ignore
+    assert 'sessionStorage.setItem("ID_TOKEN", ""' in html
+    assert 'sessionStorage.setItem("ACCESS_TOKEN", "");' in html
+    assert 'sessionStorage.setItem("REFRESH_TOKEN", "");' in html
+
+
+@pytest.mark.asyncio
+async def test_jwt_handler_get_refresh_tokens_response():
+    handler = JWTOpenIDTokensHandler(
+        JWTBearerAuthentication(
+            authority=MOCKED_AUTHORITY,
+            valid_audiences=["NULL"],
+        )
+    )
+    response = await handler.get_refresh_tokens_response(
+        Request("POST", b"/example", []),
+        TokenResponse(
+            {"id_token": "example", "access_token": "a", "refresh_token": "b"}
+        ),
+    )
+
+    assert isinstance(response.content, Content)
+    assert response.content.type == b"application/json"
+    raw_json = response.content.body  # type: ignore
+
+    assert b'"id_token":"example"' in raw_json
+    assert b'"access_token":"a"' in raw_json
+
+    # the refresh token must be protected
+    assert b'"refresh_token":"b"' not in raw_json
+    assert b'"refresh_token":"' in raw_json
+
+    data = json.loads(raw_json.decode("utf8"))
+    assert handler._serializer.loads(data["refresh_token"]) == "b"
+
+
+@pytest.mark.asyncio
+async def test_jwt_handler_restore_refresh_token():
+    handler = JWTOpenIDTokensHandler(
+        JWTBearerAuthentication(
+            authority=MOCKED_AUTHORITY,
+            valid_audiences=["NULL"],
+        )
+    )
+    context = Request(
+        "GET",
+        b"/",
+        [(b"X-Refresh-Token", handler.protect_refresh_token("Example").encode())],
+    )
+
+    handler.restore_refresh_token(context)
+    assert context.user is not None
+    assert context.user.refresh_token == "Example"
+
+
+def test_id_token_value():
+    id_token = IDToken("x", {})
+    assert id_token.value == "x"
+    assert str(id_token) == "x"